backend/requirements.txt
gradio>=4.44.0
python-dotenv
pillow
openai>=1.30.0